refactor: replace redirect-based retries with internal lock retries
